+++ R/tfidf.R
@@ -63,6 +63,11 @@
 #' }
 #' @export
 #' @details
+#' It is strongly advised to use [step_tokenfilter] before using [step_tfidf] to 
+#' limit the number of variables created, otherwise you might run into memmory
+#' issues. A good strategy is to start with a low token count and go up 
+#' according to how much RAM you want to use.
+#' 
 #' Term frequency-inverse document frequency is the product of two statistics.
 #' The term frequency (TF) and the inverse document frequency (IDF). 
 #'