+++ R/tokenmerge.R
@@ -1,12 +1,13 @@
 #'  Generate the basic set of text features
 #'
-#' `step_sequence_onehot` creates a *specification* of a recipe step that
-#'  will take a string and do one hot encoding for each character by position.
+#' `step_tokenmerge` creates a *specification* of a recipe step that
+#'  will take multiple list-columns of tokens and combine them into one 
+#'  list-column.
 #'
 #' @param recipe A recipe object. The step will be added to the
 #'  sequence of operations for this recipe.
 #' @param ... One or more selector functions to choose variables.
-#'  For `step_sequence_onehot`, this indicates the variables to be encoded
+#'  For `step_tokenmerge`, this indicates the variables to be encoded
 #'  into a list column. See [recipes::selections()] for more
 #'  details. For the `tidy` method, these are not currently used.
 #' @param role For model terms created by this step, what analysis
@@ -16,11 +17,7 @@
 #' @param columns A list of tibble results that define the
 #'  encoding. This is `NULL` until the step is trained by
 #'  [recipes::prep.recipe()].
-#' @param length A numeric, number of characters to keep before discarding. 
-#'  Defaults to 100.
-#' @param key A character vector, characters to be mapped to integers. characters 
-#'  not in the key will be encoded as 0. Defaults to `letters`.
-#' @param prefix A prefix for generated column names, default to "seq1hot".
+#' @param prefix A prefix for generated column names, default to "tokenmerge".
 #' @param skip A logical. Should the step be skipped when the
 #'  recipe is baked by [recipes::bake.recipe()]? While all
 #'  operations are baked when [recipes::prep.recipe()] is run, some
@@ -39,7 +36,8 @@
 #' data(okc_text)
 #' 
 #' okc_rec <- recipe(~ ., data = okc_text) %>%
-#'   step_sequence_onehot(essay0) 
+#'   step_tokenize(essay0, essay1) %>%
+#'   step_tokenmerge(essay0, essay1) 
 #'   
 #' okc_obj <- okc_rec %>%
 #'   prep(training = okc_text, retain = TRUE)
@@ -50,35 +48,26 @@
 #' tidy(okc_obj, number = 1)
 #' 
 #' @export
-#' @details 
-#' The string will be capped by the length argument, strings shorter then length
-#' will be padded with empty characters. The encoding will assign a integer to 
-#' each character in the key, and will encode accordingly. Characters not in the
-#' key will be encoded as 0.
 #'
 #' @importFrom recipes add_step step terms_select sel2char ellipse_check 
 #' @importFrom recipes check_type rand_id
-step_sequence_onehot <-
+step_tokenmerge <-
   function(recipe,
            ...,
            role = "predictor",
            trained = FALSE,
            columns = NULL,
-           length = 100,
-           key = letters,
-           prefix = "seq1hot",
+           prefix = "tokenmerge",
            skip = FALSE,
-           id = rand_id("sequence_onehot")
+           id = rand_id("tokenmerge")
   ) {
     add_step(
       recipe,
-      step_sequence_onehot_new(
+      step_tokenmerge_new(
         terms = ellipse_check(...),
         role = role,
         trained = trained,
         columns = columns,
-        length = length,
-        key = key,
         prefix = prefix,
         skip = skip,
         id = id
@@ -86,17 +75,15 @@ step_sequence_onehot <-
     )
   }
 
-step_sequence_onehot_new <-
-  function(terms, role, trained, columns, length, key, prefix,
+step_tokenmerge_new <-
+  function(terms, role, trained, columns, prefix,
            skip, id) {
     step(
-      subclass = "sequence_onehot",
+      subclass = "tokenmerge",
       terms = terms,
       role = role,
       trained = trained,
       columns = columns,
-      length = length,
-      key = key,
       prefix = prefix,
       skip = skip,
       id = id
@@ -104,22 +91,16 @@ step_sequence_onehot_new <-
   }
 
 #' @export
-prep.step_sequence_onehot <- function(x, training, info = NULL, ...) {
+prep.step_tokenmerge <- function(x, training, info = NULL, ...) {
   col_names <- terms_select(x$terms, info = info)
   
-  training <- factor_to_text(training, col_names)
+  check_list(training[, col_names])
   
-  check_type(training[, col_names], quant = FALSE)
-  
-  encoded_key <- char_key(x$key)
-  
-  step_sequence_onehot_new(
+  step_tokenmerge_new(
     terms = x$terms,
     role = x$role,
     trained = TRUE,
     columns = col_names,
-    length = x$length,
-    key = encoded_key,
     prefix = x$prefix,
     skip = x$skip,
     id = x$id
@@ -129,80 +110,43 @@ prep.step_sequence_onehot <- function(x, training, info = NULL, ...) {
 #' @export
 #' @importFrom tibble as_tibble
 #' @importFrom recipes bake prep
-#' @importFrom purrr map_dfc
-bake.step_sequence_onehot <- function(object, new_data, ...) {
+#' @importFrom purrr map_dfc pmap
+bake.step_tokenmerge <- function(object, new_data, ...) {
   col_names <- object$columns
   # for backward compat
+  new_col <- 
+    tibble(pmap(as.list(unname(new_data[, col_names, drop = FALSE])), c))
+  names(new_col) <- object$prefix
   
-  new_data <- factor_to_text(new_data, col_names)
+  new_data <- bind_cols(new_data, new_col)
   
-  for (i in seq_along(col_names)) {
-    out_text <- string2encoded_matrix(new_data[, col_names[i], drop = TRUE],
-                                      key = object$key, length = object$length)
-    
-    
-    colnames(out_text) <- paste(sep = "_",
-                                object$prefix, 
-                                col_names[i], 
-                                seq_len(ncol(out_text)))
-    
-    new_data <- bind_cols(new_data, as_tibble(out_text))
-    
-    new_data <-
-      new_data[, !(colnames(new_data) %in% col_names[i]), drop = FALSE]
-  }
+  new_data <-
+    new_data[, !(colnames(new_data) %in% col_names), drop = FALSE]
   
   as_tibble(new_data)
 }
 
 #' @importFrom recipes printer
 #' @export
-print.step_sequence_onehot <-
+print.step_tokenmerge <-
   function(x, width = max(20, options()$width - 30), ...) {
-    cat("Sequence 1 hot encoding for ", sep = "")
+    cat("Merging tokens for ", sep = "")
     printer(x$columns, x$terms, x$trained, width = width)
     invisible(x)
   }
 
-#' @rdname step_sequence_onehot
-#' @param x A `step_sequence_onehot` object.
+#' @rdname step_tokenmerge
+#' @param x A `step_tokenmerge` object.
 #' @importFrom recipes sel2char
 #' @export
-tidy.step_sequence_onehot <- function(x, ...) {
+tidy.step_tokenmerge <- function(x, ...) {
   if (is_trained(x)) {
     term_names <- sel2char(x$terms)
-    res <- tibble(terms = rep(term_names, each = length(x$key)), 
-                  key = rep(names(x$key), length(x$terms)))
+    res <- tibble(terms = term_names)
   } else {
     term_names <- sel2char(x$terms)
-    res <- tibble(terms = term_names,
-                  key = NA_character_)
+    res <- tibble(terms = term_names)
   }
   res$id <- x$id
   res
-}
-
-pad_string <- function(x, n) {
-  len_x <- length(x)
-  if (len_x == n) {
-    return(x)
-  }
-  c(x, character(n - len_x))
-}
-
-char_key <- function(x) {
-  out <- seq_along(x)
-  names(out) <- x
-  out
-}
-
-string2encoded_matrix <- function(x, key, length) {
-  x <- stringr::str_sub(x, 1, length)
-  x <- stringr::str_split(x, "")
-  x <- lapply(x, pad_string, n = length)
-  x <- lapply(x, function(x) key[x])
-  df <- do.call(rbind, x)
-  df[is.na(df)] <- 0
-  
-  df
 }